refactor(receiver-mock): move samples to a separate module
